studio/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
python-dotenv
langchain_ollama 
